requirements.txt
aioredis==1.3.1
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.5.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
caer==1.9.8
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.1
Django==3.2.5
django-adminlte-3==0.1.6
django-allauth==0.44.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-imagefit==0.6.2
django-js-asset==1.2.2
django-tastypie==0.14.3
django-user-unique-email==0.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
entrypoints==0.3
filelock==3.0.12
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
mypy==0.910
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
mysqlclient # Install "mysqlclient-1.4.6-cp39-cp39-win_amd64.whl" and place it in the project directory and run "pip install mysqlclient"
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
opencv-contrib-python==4.5.3.56
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.2
pipenv==2020.11.15
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2020.4
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
service-identity==21.1.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.2.4
stripe==2.57.0
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.2.1
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
